refactor(cli): centralize shared imports
