+++ src/cli/lib/shared-deps.js
@@ -3,6 +3,7 @@ export {
     assertFunction,
     assertNonEmptyString,
     assertPlainObject,
+    hasOwn,
     coerceNonNegativeInteger,
     coercePositiveInteger,
     createEnvConfiguredValue,
@@ -12,7 +13,10 @@ export {
     getOrCreateMapEntry,
     isNonEmptyArray,
     isNonEmptyString,
+    isNonEmptyTrimmedString,
+    isErrorWithCode,
     isObjectLike,
+    toMutableArray,
     normalizeIdentifierMetadataEntries,
     normalizeEnumeratedOption,
     parseJsonWithContext,
@@ -27,7 +31,9 @@ export {
 export { normalizeStringList } from "../../shared/utils/string.js";
 
 export {
+    collectAncestorDirectories,
     collectUniqueAncestorDirectories,
+    isPathInside,
     resolveContainedRelativePath
 } from "../../shared/utils/path.js";
 
@@ -39,3 +45,7 @@ export {
     isAggregateErrorLike,
     isErrorLike
 } from "../../shared/utils/capability-probes.js";
+
+export { isJsonParseError } from "../../shared/json-utils.js";
+
+export { ensureDir } from "../../shared/utils/fs.js";
